backend/requirements.txt
langchain==0.1.0
langchain-community==0.0.13
chromadb==0.4.22
sentence-transformers==2.3.1
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
huggingface-hub==0.20.2
